Replace classify with method to handle dots
The current implementation with `classify` does not work, as it does not translate dots in the name (which can be used to indicate relations). And as dots are not allowed in methods, it does not work. 

I've also added the capitalize filter such that the method still start with `formatSomething`. I would not use the classify filter anymore as it might result in conflicts with names.