Update boolean + checkbox field guesser (ORM)
Consider the following case:
dbType: boolean
formType: checkbox
Field is not nullable in the database. 

This results in a `required: true` from the field guesser, which forces the user to tick the box before anything can be saved. This PR fixes the problem, by changing the guess for this scenario to `required: false`.